+++ ricaun.Revit.UI.Tests/Resources/Images/Resources.Designer.cs
@@ -8,7 +8,7 @@
 // </auto-generated>
 //------------------------------------------------------------------------------
 
-namespace ricaun.Revit.UI.Example.Proprieties {
+namespace ricaun.Revit.UI.Tests.Resources.Images {
     using System;
     
     
@@ -22,14 +22,14 @@ namespace ricaun.Revit.UI.Example.Proprieties {
     [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "17.0.0.0")]
     [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
     [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
-    internal class Resource {
+    internal class Resources {
         
         private static global::System.Resources.ResourceManager resourceMan;
         
         private static global::System.Globalization.CultureInfo resourceCulture;
         
         [global::System.Diagnostics.CodeAnalysis.SuppressMessageAttribute("Microsoft.Performance", "CA1811:AvoidUncalledPrivateCode")]
-        internal Resource() {
+        internal Resources() {
         }
         
         /// <summary>
@@ -39,7 +39,7 @@ namespace ricaun.Revit.UI.Example.Proprieties {
         internal static global::System.Resources.ResourceManager ResourceManager {
             get {
                 if (object.ReferenceEquals(resourceMan, null)) {
-                    global::System.Resources.ResourceManager temp = new global::System.Resources.ResourceManager("ricaun.Revit.UI.Example.Proprieties.Resource", typeof(Resource).Assembly);
+                    global::System.Resources.ResourceManager temp = new global::System.Resources.ResourceManager("ricaun.Revit.UI.Tests.Resources.Images.Resources", typeof(Resources).Assembly);
                     resourceMan = temp;
                 }
                 return resourceMan;
@@ -61,11 +61,21 @@ namespace ricaun.Revit.UI.Example.Proprieties {
         }
         
         /// <summary>
+        ///   Looks up a localized resource of type System.Drawing.Icon similar to (Icon).
+        /// </summary>
+        internal static System.Drawing.Icon Revit {
+            get {
+                object obj = ResourceManager.GetObject("Revit", resourceCulture);
+                return ((System.Drawing.Icon)(obj));
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized resource of type System.Drawing.Bitmap.
         /// </summary>
-        internal static System.Drawing.Bitmap LargeImage {
+        internal static System.Drawing.Bitmap Revit32 {
             get {
-                object obj = ResourceManager.GetObject("LargeImage", resourceCulture);
+                object obj = ResourceManager.GetObject("Revit32", resourceCulture);
                 return ((System.Drawing.Bitmap)(obj));
             }
         }