+++ CHANGELOG.md
@@ -4,6 +4,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] / 2023-11-19 - 2023-11-29
+### Features
+- Update to support `net7.0-windows` and `net8.0-windows`
+- SetImage works with `Resources` without assembly name.
+- Change `System.Drawing` to a separate namespace and class.
+- Icons `GetBitmapSource` return the biggest frame.
+### Updated
+- Update `Build` project
+- Update `Example` project
+- Fix `MovePanelTo` remove panel from `RibbonTabsDictionary`
+- Add `StackTraceUtils` to find the caller assembly.
+- Update `BitmapExtension` to enable `Resources` without assembly name.
+### Tests
+- Test `MovePanelTo_Modify_CreatePanel_SameName`
+- Test `StackTraceUtils`
+- Test `Resources` without assembly name.
+- Test `Drawing` Resources.
+- Test `ResourceTests`, `ResourcePngTests` and `ResourcesFramesTests`
+
 ## [0.5.7] / 2023-11-17
 ### Features
 - LargeImage changes `TextBox` Image
@@ -301,6 +320,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: ../../compare/1.0.0...HEAD
+[0.6.0]: ../../compare/0.5.7...0.6.0
 [0.5.7]: ../../compare/0.5.6...0.5.7
 [0.5.6]: ../../compare/0.5.5...0.5.6
 [0.5.5]: ../../compare/0.5.4...0.5.5